% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_answer.R
\name{retrieve_answer}
\alias{retrieve_answer}
\title{Funcion para que el usuario digite un numero entre 1 y 6 y le retorne la solucion de uno de los ejercicios propuestos en el Classwork 1}
\usage{
retrieve_answer(seleccionar)
}
\arguments{
\item{seleccionar}{es un numero}
}
\value{
De acuerdo con la opcion seleccionada, el programa llevara al usuario al numero del ejercicio solucionado, mostrando al usuario un mensanje de la opcion seleccionada
en caso de que el numero no se encuentre dentro de los numeros del 1 al 6, arrojara un error
}
\description{
Funcion para que el usuario digite un numero entre 1 y 6 y le retorne la solucion de uno de los ejercicios propuestos en el Classwork 1
}
\examples{
retrieve_answer(1)
retrieve_answer(2)
}
